rsnaker::controls

Module input

Enums

GreetingOption

Constants

ENTER_KEYS 🔒

FRUITS_KEYS 🔒

HELP_KEYS 🔒

MAIN_MENU_KEYS 🔒

NEXT_KEYS 🔒

PARAMETERS_KEYS 🔒

PAUSE_KEYS 🔒

PREVIOUS_KEYS 🔒

QUIT_KEYS 🔒

RESET_KEYS 🔒

START_KEYS 🔒

VELOCITY_KEYS 🔒

Functions

flush_input_buffer 🔒

greeting_screen_manage_input

Check input on greeting screen Return Some(GreetingOption) if input is valid, with the chosen Greeting Option, None otherwise

playing_input_loop

You cannot block middle-click paste/scroll behavior from inside your Rust TUI app. If you really want to disable it, you would have to modify user system settings or terminal emulator config (e.g., in alacritty, kitty, gnome-terminal, etc.) That is outside the app’s control