rsnaker::controls

Module main_menu

Enums

MainMenuInput

MenuFlow 🔒

The control part of the main menu allows switching to a submenu (Fruits, Speed, Parameters, etc.) Use MainMenuInput to known which keys have been used and GreetingSimpleDisplay to display an easy menu, without input control (all except run and parameters) Return true if the player wants to play, false otherwise

Constants

ENTER_KEYS

FRUITS_KEYS 🔒

HELP_KEYS 🔒

HIGH_SCORE_KEYS 🔒

NEXT_KEYS

PARAMETERS_KEYS 🔒

PREVIOUS_KEYS

SWITCH_MENUS_OPTION 🔒

VELOCITY_KEYS 🔒

Functions

controls_main_switch_menu

enter_menu_screen 🔒

flush_input_buffer 🔒

main_menu_event

Check input on the greeting screen Return Some(GreetingOption) if input is valid, with the chosen Greeting Option, None otherwise

run_submenu_and_reset 🔒