rsnaker::controls::playing_input

Function playing_input_loop

pub fn playing_input_loop(
    direction: &Arc<RwLock<Direction>>,
    gs: &Arc<RwLock<GameState>>,
)

You cannot block middle-click paste/scroll behavior from inside your Rust TUI app. If you really want to disable it, you would have to modify user system settings or terminal emulator config (e.g., in alacrity, kitty, gnome-terminal, etc.) That is outside the app’s control

Panics

if Arc panics while holding the resources (poisoning), no recovery mechanism implemented better crash