rsnaker::game_logic::fruits_manager

Struct FruitsManager

pub struct FruitsManager<'a, 'b: 'a> {
    fruits: Vec<Fruit<'a>>,
    carte: Arc<RwLock<Map<'b>>>,
    fruit_timer_enabled: bool,
    fruit_base_lifetime: f32,
}

Manages fruit objects within the game logic. Map outlive fruits, so ’b lifetime >= ’a (fruits) lifetime

Fields
fruits: Vec<Fruit<'a>>carte: Arc<RwLock<Map<'b>>>fruit_timer_enabled: boolfruit_base_lifetime: f32
Implementations

impl<'a, 'b> FruitsManager<'a, 'b>

pub fn new( nb: u16, carte: Arc<RwLock<Map<'b>>>, fruit_timer_enabled: bool, fruit_base_lifetime: f32, ) -> Self

Creates a new FruitsManager with a given number of fruits.

Panics

if guard cannot be got for Map (whenever a previous panic poisoned guard)

Example

use std::sync::{Arc, RwLock};
use std::time::Duration;
use ratatui::layout::Rect;
use rsnaker::game_logic::fruits_manager::FruitsManager;
use rsnaker::graphics::sprites::map::Map;
let map = Arc::new(RwLock::new(Map::new(2, Rect::new(0,0, 160,10))));
let manager = FruitsManager::new(3, map, true, 5f32);

fn spawn_random( carte: &Map<'_>, fruit_timer_enabled: bool, fruit_base_lifetime: f32, ) -> Fruit<'a>

Spawns a fruit at a random position on the map.

pub fn replace_fruits(&mut self, fruits_to_remove: &[Fruit<'a>])

Replaces eaten fruits with new random ones and ensures balance.

Panics

if guard cannot be got for Map (whenever a previous panic poisoned guard)

pub fn update(&mut self)

Removes expired fruits and replaces them with fresh ones.

fn spawn_missing_fruits(&mut self, count: usize)

Refills the fruit vector to its expected capacity by spawning new fruits.

pub fn set_timer_paused(&mut self, paused: bool)

Pauses or resumes all fruit timers.

pub fn eat_some_fruits(&self, position: &Position) -> Option<Vec<Fruit<'a>>>

Returns a list of fruits at the given position, copying them to avoid lock contention.

fn generate_position_rounded_by_cs(carte: &Map<'_>) -> Position

Generates a random valid position for spawning fruits.

pub(crate) fn reset_to_terminal_size(&mut self)

pub(crate) fn reset(&mut self)

pub fn get_fruits(&self) -> &Vec<Fruit<'a>>

Returns all current fruits.

fn init(&mut self, nb: u16)