rsnaker::game_logic

Module game_options

Macros

define_args_withs 🔒

To be able to iterate over range in a meta-way, see in table parameter very useful

Structs

GameOptions

Structure holding all the configuration parameters for the game

Constants

DISPLAYABLE_EMOJI

FRUIT_DURATION_SECONDS 🔒

INI_POSITION

Initial position of the snake’s head at the start of the game

LIFE 🔒

MAX_EMOJI_BY_LINE_COUNT 🔒

NB_OF_FRUITS 🔒

ONLY_FOR_CLI_PARAMETERS

PARAMS_HEADER 🔒

PRESETS 🔒

SNAKE_GROWTH_FACTOR 🔒

SNAKE_LENGTH 🔒

define a const to avoid str errors

Functions

clamp_to 🔒

default_false 🔒

get_parameter_range

Returns the valid range for the parameter in O(1) or None

get_parameter_range_parser 🔒

Get a clap value parser for a specific parameter or the default range of 1..99

pretty 🔒