rsnaker::game_logic::game_options

Struct GameOptions

pub struct GameOptions {

    pub speed: Speed,
    pub head_symbol: String,
    pub body_symbol: String,
    pub snake_length: u16,
    pub life: u16,
    pub nb_of_fruits: u16,
    pub(crate) no_fruit_timer: bool,
    pub fruit_timer: bool,
    pub fruit_duration_seconds: u16,
    pub snake_growth_factor: u16,
    pub(crate) no_negative_size_fruits: bool,
    pub negative_size_fruits: bool,
    pub log_level: LogLevel,
    pub(crate) no_caps_fps: bool,
    pub caps_fps: bool,
    pub load: Option<u16>,
    pub random: bool,

}

Structure holding all the configuration parameters for the game

Fields
speed: Speed

Speed of the snake (Slow, Normal, Fast, Crazy) Derives ValueEnum on the enum Speed and enforces the type clap::ValueEnum, which automatically handles possible values and displays them in the help message. Now, clap enforces valid inputs without requiring a manual FromStr implementation.

head_symbol: String

Snake symbol (emoji or character) Defines short value because doublon, as short and long, are by default based on the name of the variable Default is a Christmas tree

body_symbol: String

Snake trail symbol (emoji or character) need to operate over graphene not chars see https://crates.io/crates/unicode-segmentation/ / Or deep explanation:https://docs.rs/bstr/1.12.0/bstr/#when-should-i-use-byte-strings Default is snow emoji

snake_length: u16

Initial length of the snake

life: u16

Number of lives

nb_of_fruits: u16

Number of fruits in the game

no_fruit_timer: bool

Enables the fruit timer and automatic replacement when a fruit expires.

fruit_timer: boolfruit_duration_seconds: u16

Base fruit lifetime in seconds before per-fruit multipliers are applied.

snake_growth_factor: u16

Multiplier applied to the hardcoded snake-size effect of each fruit.

no_negative_size_fruits: bool

Allow fruits with negative size effects.

negative_size_fruits: boollog_level: LogLevel

Logging level

no_caps_fps: bool

Modern way to do CLI, two dedicated flag to set/unset the value, beginning with –no- (for false) UX better than –feature false / –feature true, better than default (no flag = false). If you want the possibility to set both values, as no clear default value or want to be able to easily programmatically change the value (as there) or to have a default at true
See: https://jwodder.github.io/kbits/posts/clap-bool-negate/ As default is true, no and value are swaped

caps_fps: boolload: Option<u16>

Load game parameters

random: bool

Randomize all game parameters with playable values

Implementations

impl GameOptions

pub fn initial_position() -> Position

Returns the initial snake position

pub fn emojis_with_news_line() -> String

pub fn emojis_iterator() -> impl Iterator<Item = String>

pub fn to_structured_toml(&self) -> Table

To be editable easily

Panics

if self cannot be parsed (not possible)

pub fn validate(&mut self)

Validates and clamps the parameters to their allowed ranges. Ensures that numeric values are within the bounds defined in the CLI macro and that symbols are exactly one grapheme long. Enums are already checked automatically during deserialization

pub fn load_from_toml_preset(preset: u16) -> Result<Self>

Load parameters from a preset TOML configuration

Errors

Returns an error if the preset file cannot be opened or read, or if it contains invalid TOML.

fn load_from_toml<P: AsRef<Path>>(path: P) -> Result<Self>

Load parameters from a TOML file

Errors

Returns an error if the file cannot be opened or read.

Panics

Panic if the file contents cannot be deserialized as valid TOML.

pub fn save_to_toml_preset(&mut self, preset: u16) -> Result<()>

Save parameters to a preset TOML configuration

Errors

Returns an error if the preset file cannot be created or written to.

fn save_to_toml<P: AsRef<Path>>(&mut self, path: P) -> Result<()>

Save the current parameters to a TOML file NB: non-serializable parameters are marked skip with serde annotation and will not be included

Errors

Returns an error if the file cannot be created or written to.

Panics

Panics if the game parameters cannot be serialized to TOML.

pub fn random() -> Self

Returns a new GameOptions with randomized playable parameters using weighted distributions.