rsnaker::game_logic

Module playing_thread_manager

Re-exports

pub use crate::controls::direction::Direction;

pub use crate::controls::speed::Speed;

pub use crate::game_logic::fruits_manager::FruitsManager;

pub use crate::game_logic::state::GameState;

pub use crate::game_logic::state::GameStatus;

Structs

Game

our game engine NB: ’c must outlive ’b as, ’c (fruits manager) uses in intern the map with lock on it. NB: ’t the terminal life must outlive all the other lifetimes