rsnaker::graphics::menus

Module main_menu

Structs

Button

Represents a button in the menu interface

Enums

SwitchMenu

Option that can be selected from the main menu, using a lateral switcher

Constants

BUTTONS 🔒

Statics

EDGE_SNAKE 🔒

Functions

big_snake_menu 🔒

display_main_menu

Print the wanted welcome screen controls Show Fruit and Speed menus alongside Sadly slow_blink and fast_blink are not rendered anymore on modern terminal…

get_button_span 🔒

Returns a vector of spans representing the button navigation menu