rsnaker::graphics

Module playing_render

Constants

BOTTOM_SPEED_FPS_SCORE_RECT 🔒

Position to render elements will be clamped to the frame area border anyway, 9999 to go to the last line allowing easy resizing

LIFE_RECT 🔒

NB_OF_FRAMES_WINDOW 🔒

TOO_MUCH_LIVES_TO_DISPLAY 🔒

Functions

game_state_render 🔒

Return whether stop the rendering

playing_render_loop

Panics